requirements.txt
fastapi==0.70.1
uvicorn==0.16.0
pydantic==1.8.2
requests==2.28.1
python-dotenv==0.19.2
